Fix integer overflow in nccl thunk

+++ tensorflow/compiler/xla/service/gpu/nccl_collective_permute_thunk.cc
@@ -179,7 +179,7 @@ Status NcclCollectivePermuteThunk::RunNcclCollective(
   TF_ASSIGN_OR_RETURN(auto dtype_and_multiplier,
                       ToNcclDataTypeAndCountMultiplier(element_type));
   ncclDataType_t dtype = dtype_and_multiplier.first;
-  int element_count = buffer_.element_count * dtype_and_multiplier.second;
+  int64_t element_count = buffer_.element_count * dtype_and_multiplier.second;
 
   cudaStream_t* cu_stream = reinterpret_cast<cudaStream_t*>(
       params.stream->implementation()->GpuStreamMemberHack());
